Login (Features): -Now it is possible (again) to enter in offline mode and visit the homepage and services available; -Added a option to Register users, Registering users opens a webpage on your browser and it's still not possible to register users inside the app.
MainPage/HomePage (Features):
Added option to view the owner's vehicles;
Added options for the users to check when will be the next schedule service to the car.

Login (Visual): Small tweaks

MainPage/HomePage (Visual):
Added a image background to the page;
Added listviews and different framelayouts for the vehicles and schedules activities.

ATTENTION: There is still more work to be done on the framelayout contents as well as the offline mode.